Pass all packages in at once
Since it's much more efficient to pass all packages into the module at once
rather than using a loop, create separate tasks for packages with and without
additional options.

Originally this was combined into a single task to avoid the need for two
separate lists of packages to install. But the performance hit for looping
over every item in the list outweighs the minor inconvenienc of having to
specify multiple lists.