🐛 Keep just one DMG in the auto-deploy package (#15)
The official `pd-autodeploy.zip` has a bug. Its `postflight` script
selects an image to install from the `*.dmg` files present under its
`Parallels Desktop DMG/` directory. But it does this by running the
`ls` command without any additional sorting. This means that the list
of files returned is sorted by name. It then takes the first entry
from that list which usually ends up being the oldest DMG in the list.
This is not a big deal if there is a single DMG present in that folder
but the way the `parallels` role was implemented before this commit
ends up allows for multiple DMG files to be accumulated in that
folder. This is where the problem manifests itself — subsequent
autodeploy invocations put a new DMG file in place but the autodeploy
script keeps selecting the old one. As a result, said autodeploy
reports installing the Parallels update successfully, but the version
does not change, despite the reported success (I suppose it could also
end up being downgraded silently in some cases).

This patch is trying to combat the issue by cleaning up all the DMGs
from the `Parallels Desktop DMG/` folder, followed by symlinking the
target one.