🔥 Rm target `.pkg` installer before making it (#56)
In the recent versions of `pd-autodeploy.zip`, there is a `prepare`
script that is supposed to be called to produce a flat `.pkg`
installer package.

Prior to this change, it would get skipped on reruns, resulting in the
following steps attempting to install from an older bundled DMG.

This change makes it disappear so it always runs. In the future, this
could be improved by writing the output to hash-based dirs.