🔥 Find DMG links @ making parallels installer (#55)
Prior to this patch, the task would only match regular files and
ignore any symlinks. This would cause the following steps to get it
into the `.pkg` and the installer would often attempt installing the
oldest DMG present, while we usually want to install something newer.

This change fixes that by making the lookup check if anything matching
is in the directory so it can be deleted before producing that `.pkg`
installer.